requirements.txt
Flask==3.0.2
tensorflow==2.15.0
opencv-python-headless==4.8.0.74
numpy==1.23.5
Pillow==10.0.1